+++ CHANGELOG.md
@@ -1,4 +1,4 @@
 Changelog for 6.0
 =================
 
-*   Added new form option `row_attr`, that acts like `attr` but is dedicated for form row attributes.
+*   Added new form option `attr_row`, that acts like `attr` but is dedicated for form row attributes.